tests/test_ticketing.py
```python
import unittest
from src import ticketing

class TestTicketing(unittest.TestCase):

    def setUp(self):
        self.ticketing_system = ticketing.TicketingSystem()
        self.user_data = {
            "username": "test_user",
            "email": "test_user@example.com"
        }
        self.event_data = {
            "event_name": "Test Event",
            "event_date": "2022-12-31",
            "event_time": "12:00:00"
        }

    def test_purchase_ticket(self):
        # Test if a ticket can be purchased successfully
        purchase_status = self.ticketing_system.purchase_ticket(self.user_data, self.event_data)
        self.assertEqual(purchase_status, "Ticket purchased successfully")

    def test_ticket_confirmation(self):
        # Test if a ticket confirmation is sent successfully
        confirmation_status = self.ticketing_system.send_ticket_confirmation(self.user_data, self.event_data)
        self.assertEqual(confirmation_status, "Ticket confirmation sent successfully")

    def test_ticket_availability(self):
        # Test if the ticket availability status is correct
        availability_status = self.ticketing_system.check_ticket_availability(self.event_data)
        self.assertTrue(availability_status)

if __name__ == '__main__':
    unittest.main()
```